% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dharma.r
\name{.compute_dharma}
\alias{.compute_dharma}
\title{Compute DHARMa residuals for model diagnostics}
\usage{
.compute_dharma(x, nsim)
}
\arguments{
\item{x}{An immuno_fit object}

\item{nsim}{Number of simulations for DHARMa}
}
\value{
DHARMa object or NULL if computation fails
}
\description{
Compute DHARMa residuals for model diagnostics
}
\keyword{internal}
