% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dharma.r
\name{.compute_dharma_external_censreg}
\alias{.compute_dharma_external_censreg}
\title{External DHARMa simulation for censReg models}
\usage{
.compute_dharma_external_censreg(x, nsim)
}
\arguments{
\item{x}{An immuno_fit object with censReg model}

\item{nsim}{Number of simulations}
}
\value{
DHARMa object or NULL if computation fails
}
\description{
External DHARMa simulation for censReg models
}
\keyword{internal}
