% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.r
\name{.create_basic_plots}
\alias{.create_basic_plots}
\title{Create basic diagnostic plots}
\usage{
.create_basic_plots(plot_data, plot_context, point_size, alpha)
}
\arguments{
\item{plot_data}{Data frame with plotting data}

\item{plot_context}{List with family, estimand, resid_type info}

\item{point_size}{Size of points in plots}

\item{alpha}{Transparency of points}
}
\value{
List with rvf and qq plots
}
\description{
Create basic diagnostic plots
}
\keyword{internal}
