% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.r
\name{.create_censor_aware_plots}
\alias{.create_censor_aware_plots}
\title{Create censoring-aware diagnostic plots}
\usage{
.create_censor_aware_plots(plot_data, plot_context, point_size, alpha_cens)
}
\arguments{
\item{plot_data}{Data frame with plotting data}

\item{plot_context}{List with family, estimand, resid_type info}

\item{point_size}{Size of points in plots}

\item{alpha_cens}{Transparency of censored points}
}
\value{
List with rvf and qq plots
}
\description{
Create censoring-aware diagnostic plots
}
\keyword{internal}
