% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residuals.r
\name{.get_residuals_and_fitted}
\alias{.get_residuals_and_fitted}
\title{Get residuals and fitted values for plotting}
\usage{
.get_residuals_and_fitted(x, plot_data)
}
\arguments{
\item{x}{An immuno_fit object}

\item{plot_data}{Data frame with plotting data}
}
\value{
List with residuals, fitted values, and residual type
}
\description{
Get residuals and fitted values for plotting
}
\keyword{internal}
