% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.immuno_preprocess.r
\name{plot.immuno_preprocess}
\alias{plot.immuno_preprocess}
\title{Plot percent-censored profile for immunoPlex objects}
\usage{
\method{plot}{immuno_preprocess}(x, ...)
}
\arguments{
\item{x}{An object of class \code{immuno_preprocess} or
\code{immuno_preprocess_compare}.}

\item{...}{Ignored; included for S3 consistency.}
}
\description{
These S3 methods extend \strong{graphics::plot()} so that a simple
\code{plot(obj)} call yields a bar chart of the proportion
of censored (<LOD) observations for each cytokine - no faceting by method.
}
