% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plsda_model.r
\name{plot_pls_loadings}
\alias{plot_pls_loadings}
\title{Create PLS-DA loading plot}
\usage{
plot_pls_loadings(model, components = c(1, 2), top_n = NULL, ...)
}
\arguments{
\item{model}{A plsda_model object}

\item{components}{Numeric vector of component indices to plot}

\item{top_n}{Number of top cytokines to show (by absolute loading)}

\item{...}{Additional arguments}
}
\value{
A ggplot object
}
\description{
Bar plot showing cytokine loadings on PLS-DA components.
}
