% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-plsda_model.r
\name{plot_pls_vip}
\alias{plot_pls_vip}
\title{Create VIP score plot}
\usage{
plot_pls_vip(model, vip_threshold = 1, top_n = NULL, ...)
}
\arguments{
\item{model}{A plsda_model object}

\item{vip_threshold}{Threshold for highlighting important variables}

\item{top_n}{Number of top cytokines to show}

\item{...}{Additional arguments}
}
\value{
A ggplot object
}
\description{
Horizontal bar plot of Variable Importance in Projection scores.
}
