% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_fit.r
\name{predict.plsda_model}
\alias{predict.plsda_model}
\title{Extract model predictions from plsda_model}
\usage{
\method{predict}{plsda_model}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A plsda_model object}

\item{newdata}{Optional new data to predict (must be preprocessed same way)}

\item{...}{Additional arguments (not used)}
}
\value{
A vector of predicted class labels
}
\description{
Extract model predictions from plsda_model
}
