% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.immuno_ancova.r
\name{print.immuno_ancova_set}
\alias{print.immuno_ancova_set}
\title{Print method for immuno_ancova_set}
\usage{
\method{print}{immuno_ancova_set}(x, ...)
}
\arguments{
\item{x}{An \code{immuno_ancova_set} object.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Summary counts for a batch ANCOVA result set.
}
