% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.immuno_fit.r
\name{print.immuno_fit}
\alias{print.immuno_fit}
\title{Print method for immuno_fit}
\usage{
\method{print}{immuno_fit}(x, ...)
}
\arguments{
\item{x}{An \code{immuno_fit} object.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Shows key information about the fitted model.
}
