% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.immuno_model_set.r
\name{print.immuno_lod_comparison}
\alias{print.immuno_lod_comparison}
\title{Print method for immuno_lod_comparison}
\usage{
\method{print}{immuno_lod_comparison}(x, ...)
}
\arguments{
\item{x}{An \code{immuno_lod_comparison} object}

\item{...}{Unused}
}
\value{
\code{x} invisibly
}
\description{
Shows LOD comparison results
}
