% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.immuno_model.r
\name{print.immuno_model}
\alias{print.immuno_model}
\title{Print method for immuno_model}
\usage{
\method{print}{immuno_model}(x, ...)
}
\arguments{
\item{x}{An \code{immuno_model} object.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Minimal representation for interactive use.
}
