% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.immuno_model_set.r
\name{print.immuno_model_set}
\alias{print.immuno_model_set}
\title{Print method for immuno_model_set}
\usage{
\method{print}{immuno_model_set}(x, ...)
}
\arguments{
\item{x}{An \code{immuno_model_set} object}

\item{...}{Unused}
}
\value{
\code{x} invisibly
}
\description{
Shows model comparison table and highlights best model
}
