% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_immunoassay.r
\name{print.immuno_sim}
\alias{print.immuno_sim}
\title{Print method for immuno_sim}
\usage{
\method{print}{immuno_sim}(x, ...)
}
\arguments{
\item{x}{An \code{immuno_sim} object.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
One-line summary of a simulated immunoassay dataset.
}
