% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemar_detection.r
\name{print.mcnemar_detection}
\alias{print.mcnemar_detection}
\title{Print method for McNemar detection results}
\usage{
\method{print}{mcnemar_detection}(x, n_show = 10, ...)
}
\arguments{
\item{x}{Result from mcnemar_detection()}

\item{n_show}{Number of top results to show (default: 10)}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for McNemar detection results
}
