% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_fit.r
\name{print.plsda_model}
\alias{print.plsda_model}
\title{Print method for plsda_model objects}
\usage{
\method{print}{plsda_model}(x, ...)
}
\arguments{
\item{x}{A plsda_model object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for plsda_model objects
}
