% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_preprocess.r
\name{print.plsda_preprocessed}
\alias{print.plsda_preprocessed}
\title{Print method for plsda_preprocessed objects}
\usage{
\method{print}{plsda_preprocessed}(x, ...)
}
\arguments{
\item{x}{A plsda_preprocessed object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for plsda_preprocessed objects
}
