% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{reset_immunoplex_config}
\alias{reset_immunoplex_config}
\title{Reset immunoPlex configuration to defaults}
\usage{
reset_immunoplex_config()
}
\description{
Resets all immunoPlex configuration parameters to their default values.
}
\examples{
# After experimenting with settings
reset_immunoplex_config()

}
