% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_utils.R
\name{simulate_pi_coverage}
\alias{simulate_pi_coverage}
\title{Predictive interval coverage and mean width via simulation}
\usage{
simulate_pi_coverage(fit, nsim = 200L, level = 0.95)
}
\arguments{
\item{fit}{an \code{immuno_fit} object}

\item{nsim}{integer number of simulations}

\item{level}{numeric interval level (e.g., 0.95)}
}
\description{
Predictive interval coverage and mean width via simulation
}
