% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.immuno_model_set.r
\name{summary.immuno_lod_comparison}
\alias{summary.immuno_lod_comparison}
\title{Summary method for immuno_lod_comparison}
\usage{
\method{summary}{immuno_lod_comparison}(object, ...)
}
\arguments{
\item{object}{An \code{immuno_lod_comparison} object}

\item{...}{Unused}
}
\value{
Summary object
}
\description{
Detailed LOD comparison summary
}
