% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.immuno_model_set.r
\name{summary.immuno_model_set}
\alias{summary.immuno_model_set}
\title{Summary method for immuno_model_set}
\usage{
\method{summary}{immuno_model_set}(object, ...)
}
\arguments{
\item{object}{An \code{immuno_model_set} object}

\item{...}{Unused}
}
\value{
Summary object
}
\description{
Detailed summary of all fitted models
}
