% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_fit.r
\name{summary.plsda_model}
\alias{summary.plsda_model}
\title{Summary method for plsda_model objects}
\usage{
\method{summary}{plsda_model}(object, ...)
}
\arguments{
\item{object}{A plsda_model object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for plsda_model objects
}
