% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_preprocess.r
\name{summary.plsda_preprocessed}
\alias{summary.plsda_preprocessed}
\title{Summary method for plsda_preprocessed objects}
\usage{
\method{summary}{plsda_preprocessed}(object, ...)
}
\arguments{
\item{object}{A plsda_preprocessed object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for plsda_preprocessed objects
}
